+++ spec/generators/install_generator_spec.rb
@@ -42,8 +42,8 @@ describe Logidze::Generators::InstallGenerator, type: :generator do
 
         is_expected.to be_a_file
         is_expected.to contain "ActiveRecord::Migration[#{ar_version}]"
-        is_expected.to contain("create_function :logidze_logger, version: 5")
-        is_expected.to contain("create_function :logidze_logger_after, version: 5")
+        is_expected.to contain("create_function :logidze_logger, version: 6")
+        is_expected.to contain("create_function :logidze_logger_after, version: 6")
         is_expected.to contain("create_function :logidze_snapshot, version: 3")
         is_expected.to contain("create_function :logidze_version, version: 2")
         is_expected.to contain("create_function :logidze_filter_keys, version: 1")
@@ -56,8 +56,8 @@ describe Logidze::Generators::InstallGenerator, type: :generator do
 
         is_expected.to be_a_file
         %w[
-          logidze_logger_v05.sql
-          logidze_logger_after_v05.sql
+          logidze_logger_v06.sql
+          logidze_logger_after_v06.sql
           logidze_version_v02.sql
           logidze_filter_keys_v01.sql
           logidze_snapshot_v03.sql
@@ -146,10 +146,10 @@ describe Logidze::Generators::InstallGenerator, type: :generator do
         is_expected.not_to contain("create_function :logidze_filter_keys")
         is_expected.to contain("create_function :logidze_compact_history, version: 1")
 
-        is_expected.to contain("create_function :logidze_logger, version: 5")
+        is_expected.to contain("create_function :logidze_logger, version: 6")
         is_expected.to contain("create_function :logidze_logger, version: 7")
 
-        is_expected.to contain("create_function :logidze_logger_after, version: 5")
+        is_expected.to contain("create_function :logidze_logger_after, version: 6")
         is_expected.to contain("create_function :logidze_logger_after, version: 7")
 
         is_expected.to contain("create_function :logidze_capture_exception")
@@ -160,8 +160,8 @@ describe Logidze::Generators::InstallGenerator, type: :generator do
 
         is_expected.to be_a_file
         %w[
-          logidze_logger_v05.sql
-          logidze_logger_after_v05.sql
+          logidze_logger_v06.sql
+          logidze_logger_after_v06.sql
           logidze_version_v02.sql
           logidze_snapshot_v03.sql
           logidze_compact_history_v01.sql